requirements.txt
asgiref==3.8.1
Django==3.1.13
django-lexorank @ git+https://github.com/HuyNguyenDinh/django-lexorank.git@nightly
djangorestframework==3.12.4
sqlparse==0.5.3